server

+++ server-linuxkernel-beaglebone.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
+if [ "${1}" == "install" ]; then
 sudo apt-get -y update || echo "some error"
 #sudo apt-get install flex
 sudo apt autoremove || echo "some error"
+sudo apt-get install libssl-dev
 sudo apt-get install lzma lzop bison libncurses5-dev:amd64 || echo "some error"
+elif [ "${1}" == "remove" ]; then
+sudo apt-get remove flex
+sudo apt autoremove || echo "some error"
+sudo apt-get remove libssl-dev
+sudo apt-get remove lzma lzop bison libncurses5-dev:amd64 || echo "some error"
+else
 #wget -c https://jaist.dl.sourceforge.net/project/flexformatter/flexformatter/0.9.0/FlexPrettyPrintCommand_0.9.0.zip
 #wget -c http://down1.chinaunix.net/distfiles/flex-2.5.4a.tar.gz
 #vim /etc/apt/source.list
@@ -20,3 +28,4 @@ cd bb-kernel/
 #git checkout origin/am33x-rt-v4.14 -b tmp
 git checkout origin/am33x-v4.15 -b tmp || git add .
 ./build_kernel.sh
+fi
